requirements.txt
flask-heroku
flask_cors
pandas
requests
gunicorn
whitenoise
bs4
lxml
jsonify
